% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_stage_bootstrap.R
\name{boot_stage_one}
\alias{boot_stage_one}
\title{First stage of two-stage bootstrap}
\usage{
boot_stage_one(x, group, frequency = NULL, draws)
}
\arguments{
\item{x}{A vector of the variable to be sampled.}

\item{group}{A vector of grouping variables for x.}

\item{frequency}{A numeric vector giving the frequency of x. Used for probabalistic sample draws. Default = NULL assigns equal sampling probability to each item x.}

\item{draws}{Number of group samples to draw}
}
\value{
A list containing a list of samples and associated group from which the samples were drawn.
}
\description{
Draw from paired treatment blocks and samples from first treatment level
}
