% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_stage_bootstrap.R
\name{boot_stage_two}
\alias{boot_stage_two}
\title{Bootstrap draw with resampling}
\usage{
boot_stage_two(x, frequency = NULL, group, draw_group, draw_index)
}
\arguments{
\item{x}{A vector of the variable to be sampled}

\item{frequency}{A numeric vector giving the frequency of x. Used for probabilistic sample draws. Default = NULL assigns equal sampling probability to each item x.}

\item{group}{A vector of grouping variables for x.}

\item{draw_group}{Number of group samples to draw}

\item{draw_index}{Index value for bootstrap draw (i.e., samples drawn the first sample = 1, second sample = 2, etc.)}
}
\value{
A list containing a list of samples and associated group from which the samples were drawn.
}
\description{
Bootstrap draw with resampling
}
