% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc_bias}
\alias{calc_bias}
\title{Calculate Bias}
\usage{
calc_bias(est, obs, const = 0, type = mean)
}
\arguments{
\item{est}{A numeric vector of estimated values.}

\item{obs}{A numeric vector of observed values.}

\item{const}{A numeric constant added to the estimated and observed values before calculation. Default is 0.}

\item{type}{A function to summarize the bias, such as \code{mean} or \code{median}. Default is \code{mean}.}
}
\value{
A numeric value representing the bias.
}
\description{
This function calculates the bias between estimated and observed values.
}
\examples{
est <- c(1, 2, 3)
obs <- c(1.1, 2.1, 2.9)
calc_bias(est, obs)
}
