% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gapindex.R
\name{calc_gapindex}
\alias{calc_gapindex}
\title{Function to calculate design-based products from gapindex tables and age-length keys}
\usage{
calc_gapindex(
  racebase_tables,
  alk = NULL,
  subset_fields = FALSE,
  fill_NA_method = "AIGOA"
)
}
\arguments{
\item{racebase_tables}{data object created from gapindex::get_data()}

\item{alk}{age-length key (dataframe) created from gapindex::calc_alk()}

\item{subset_fields}{Should all of the fields be returned for biomass, agecomp, and sizecomp data frames?}

\item{fill_NA_method}{passed to gapindex::calc_sizecomp_stratum()}
}
\description{
Wrapper around gapindex functions to calcualte survey biomass, abundance, age comps, and size comps.
}
