% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc_mae}
\alias{calc_mae}
\title{Calculate Mean Absolute Error (MAE)}
\usage{
calc_mae(est, obs, const = 0, type = mean)
}
\arguments{
\item{est}{A numeric vector of estimated values.}

\item{obs}{A numeric vector of observed values.}

\item{const}{A numeric constant added to the estimated and observed values before calculation. Default is 0.}

\item{type}{A function to summarize the error, such as \code{mean} or \code{median}. Default is \code{mean}.}
}
\value{
A numeric value representing the MAE.
}
\description{
This function calculates the mean absolute error between estimated and observed values.
}
\examples{
est <- c(1, 2, 3)
obs <- c(1.1, 2.1, 2.9)
calc_mae(est, obs)
}
