% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc_mre}
\alias{calc_mre}
\title{Calculate Mean Relative Error (MRE)}
\usage{
calc_mre(est, obs, type = mean)
}
\arguments{
\item{est}{A numeric vector of estimated values.}

\item{obs}{A numeric vector of observed values.}

\item{type}{A function to summarize the error, such as \code{mean} or \code{median}. Default is \code{mean}.}
}
\value{
A numeric value representing the MRE.
}
\description{
This function calculates the mean relative error between estimated and observed values.
}
\examples{
est <- c(1, 2, 3)
obs <- c(1.1, 2.1, 2.9)
calc_mre(est, obs)
}
