% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_performance_metrics.R
\name{calculate_performance_metrics}
\alias{calculate_performance_metrics}
\title{Calculate RMSLE, MRE, MAE, R Sq., Bias}
\usage{
calculate_performance_metrics(x, y, add_for_log = 1)
}
\arguments{
\item{x}{numeric vector for the first series}

\item{y}{numeric vector for the second series}

\item{add_for_log}{Value to add to x and y that require log transformations.}
}
\description{
Compare series x to y using performance metrics from Seegers et al. (2018). Root mean square log error, mean relative error, mean absolute error, R-squared, and mean bias.
}
