% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_at_size.R
\name{catch_at_size}
\alias{catch_at_size}
\title{Simulate catch using a Dirichlet-multinomial distribution}
\usage{
catch_at_size(size, abundance, selectivity, n_size_samples)
}
\arguments{
\item{size}{Vector of sizes}

\item{abundance}{Vector of numbers-at-size}

\item{n_size_samples}{Number of individuals captured per sample}

\item{selectivity_opts}{A vector of selectivity-at-size}
}
\description{
Simulate catch-at-size as the product of selectivity and numbers-at-length using a Dirichlet-multinomial distribution.
}
