% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crab_size_1995_1998}
\alias{crab_size_1995_1998}
\title{Crab size from 15/30 minute tows in 1995 and 1998}
\format{
A data frame with 12 variables:
  \describe{
    \item{HAULJOIN}{A unique identifier for each CRUISE/VESSEL/HAUL combination.}
    \item{CRUISE}{Six-digit number identifying the cruise, where the first four digits are the year of the cruise and last four digits are region-specific sequential codes.}
    \item{VESSEL}{Unique numeric vessel code.}
    \item{HAUL}{Sequential haul number identifying a unique haul for a VESSEL/CRUISE combination}
    \item{SPECIES_CODE}{Numeric species code}
    \item{SEX}{Sex, where 1 = male, 2 = female, 3 = unknown/unspecified}
    \item{LENGTH}{Carapace length in millimeters}
    \item{WIDTH}{Carapace width in millimeters}
    \item{SHELL_CONDITION}{Shell condition}
    \item{SAMPLING_FACTOR}{Sampling factor}
    \item{CHELA_HEIGHT}{Chela height in millimeters}
    \item{STATIONID}{Survey station name}
  }
}
\usage{
crab_size_1995_1998
}
\description{
Crab size measurement data from paired 15 and 30 minute tows conducted in the eastern Bering Sea in 1995 and 1998 aboard the FV Arcturus and FV Aldebaran. Sampling protocols and methods are described in Goddard et al. (1997) and Somerton et al. (2002).
}
\references{
Goddard, P.D., 1997. The effects of tow duration and subsampling on CPUE, species composition and length distributions of bottom trawl survey catches. Master of Science Thesis, University of Washington, Seattle, WA, 119 pp.
 Somerton, D.A., Otto, R.S., and Syrjala, S.E. 2002. Can changes in tow duration on bottom trawl surveys lead to changes in CPUE and mean size? Fisheries Research, 55(1–3), 63–70. https://doi.org/10.1016/S0165-7836(01)00293-4
}
\keyword{datasets}
