% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_1530}
\alias{data_1530}
\title{15/30 Comparison Data}
\format{
An object of class 'list'
}
\usage{
data_1530
}
\description{
List containing haul, catch, length, and crab size data for the 15/30 comparison project.
}
\keyword{datasets}
