% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_ss}
\alias{data_ss}
\title{Shelf/slope Comparison Data}
\format{
An object of class 'list'
}
\usage{
data_ss
}
\description{
List containing haul, catch, and size (fish length, crab carapace width, crab carapace length) for the shelf/slope comparison project.
}
\keyword{datasets}
