% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_fit_bootstrap.R
\name{discrete_fit_bootstrap}
\alias{discrete_fit_bootstrap}
\title{Bootstrap selectivity ratio}
\usage{
discrete_fit_bootstrap(
  x,
  treatment_order,
  size_col,
  block_col,
  treatment_col,
  count_col,
  effort_col,
  gam_family,
  scale_method = "sv",
  n_cores = 1
)
}
\arguments{
\item{x}{list containing bootstrap samples, output from two_stage_bootstrap}

\item{treatment_order}{rank order of treatment levels, must match treatment names in treatment col}

\item{size_col}{Character vector; name of the size/age column}

\item{block_col}{Character vector; name of the sampling block column}

\item{treatment_col}{Character vector; name of the column containing treatment names}

\item{count_col}{Character vector; name of the column containing counts}

\item{effort_col}{Character vector; name of the column containing effort}

\item{gam_family}{"binomial" or "beta"}

\item{scale_method}{Method to use for scaling the catch comparison rate for beta regression. See ?scale_for_betareg}

\item{n_cores}{Number of cores to use for parallel processing.}
}
\description{
Uses the output of two_stage_bootstrap to fit discrete size/age binomial or beta generalized additive mixed models for catch comparison rate to generate selectivity ratio estimates.
}
