% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_norm_24.R
\name{double_norm_24}
\alias{double_norm_24}
\title{Double normal selectivity function}
\usage{
double_norm_24(x, a, b, c, d, e, f)
}
\arguments{
\item{x}{Vector of lengths}

\item{a}{Parameter a}

\item{b}{Parameter b}

\item{c}{Parameter c}

\item{d}{Parameter d}

\item{e}{Parameter e}

\item{f}{Parameter f}
}
\description{
Selectivity pattern 24 in SS3. Copied from https://github.com/r4ss/r4ss/blob/414f60b88f404a92eb7eab323c6cc580798565ea/inst/Shiny/Selectivities/server.R on October 26, 2021.
}
