% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_from_dist.R
\name{draw_from_dist}
\alias{draw_from_dist}
\title{Draw samples from a distribution}
\usage{
draw_from_dist(x)
}
\arguments{
\item{x}{A list containing arguments passed to the distribution of choice.}
}
\description{
Draw samples from a distribution
}
\details{
x should be a list containing an argument named distribution denoting the distribution to use (e.g. 'normal') and parameters passed to the distributional function for the distribution (e.g. the normal distribution specifying the use of rnorm should have mean and sd arguments)
}
