% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gapindex_1530}
\alias{gapindex_1530}
\title{gapindex data for the 15/30 comparison}
\format{
An object of class 'list'
}
\usage{
gapindex_1530
}
\description{
List containing RACEBASE tables created using the gapindex package for the 15/30 comparison project.
}
\keyword{datasets}
