% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_connected}
\alias{get_connected}
\title{Connect to Oracle}
\usage{
get_connected(schema = "AFSC")
}
\arguments{
\item{schema}{Connection. Default = 'AFSC'}
}
\description{
Connect to set up a connection to Oracle using RODBC and getPass
}
