% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_population_size_comp.R
\name{make_population_size_comp}
\alias{make_population_size_comp}
\title{Simulate population size composition from mean size at age}
\usage{
make_population_size_comp(
  age,
  size,
  abundance,
  sd_at_age_df,
  size_range = c(3, 127)
)
}
\arguments{
\item{age}{Vector of age}

\item{size}{Vector of mean length at age}

\item{abundance}{Vector of abundances at size}

\item{sd_at_age_df}{Data frame containing standard deviation at age 1 and 20 for the chosen ensemble.}

\item{size_range}{Numeric vector of 2L. Maximum and minimum lengths}
}
\description{
Simulate size composition of the population based on numbers-at-age, size-at-age, and standard deviation in size-at-age.
}
