% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prediction_df.R
\name{make_prediction_df}
\alias{make_prediction_df}
\title{Make GAM prediction data.frame}
\usage{
make_prediction_df(model, lengths = 10:120, type = "link", model_name)
}
\arguments{
\item{model}{GAM model}

\item{lengths}{lengths to predict}

\item{type}{type of prediction to return}

\item{model_name}{name/type of model used to generate the prediction}
}
\description{
Make GAM prediction data.frame
}
