% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_catch_bootstrap.R
\name{nested_bootstrap}
\alias{nested_bootstrap}
\title{Bootstrap draw with resampling}
\usage{
nested_bootstrap(x, frequency = NULL, group, draw_group, draw_index)
}
\arguments{
\item{x}{A vector of the variable to be sampled}

\item{frequency}{A numeric vector giving the frequency of x. Used for probabilistic sample draws. Default = NULL assigns equal sampling probability to each item x.}

\item{group}{A vector of grouping variables for x.}

\item{draw_group}{Number of group samples to draw}
}
\value{
A list containing a list of samples and associated group from which the samples were drawn.
}
\description{
Bootstrap draw with resampling
}
