% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otto_key_1998}
\alias{otto_key_1998}
\title{Blocks of 15/30 minute tows from 1998}
\format{
A data frame with 6 variables:
  \describe{
    \item{TOW_BLOCK}{Value identifying the *block* to which a tow was assigned.}
    \item{TOW_PAIR}{Value identifying the tow *pair* for this data set.}
    \item{EXPERIMENTAL_TOW_TYPE}{Numerical tow type code.}
    \item{CRUISE}{Six-digit number identifying the cruise, where the first four digits are the year of the cruise and last four digits are region-specific sequential codes.}
    \item{VESSEL}{Unique numeric vessel code.}
    \item{HAUL}{Sequential haul number identifying a unique haul for a VESSEL/CRUISE combination}
    \item{TOW_TYPE}{Description of numberical tow type}
  }
}
\usage{
data(otto_key_1998)
}
\description{
Key identifying blocks of 15/30 minute tows adapted from Bob Otto's tow pair spreadsheet (2000).
}
\keyword{datasets}
