% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bootstrap.R
\name{run_bootstrap}
\alias{run_bootstrap}
\title{Run bootstrap sampling for N iterations}
\usage{
run_bootstrap(dat, lengths, iterations = 1000, clusters = 2, seed, k = 10)
}
\arguments{
\item{dat}{data.frame containing a columns named p_fishery_adj, MATCHUP, LEN_MIDPOINT, dummy.}

\item{lengths}{numeric vector of fork lengths for prediction}

\item{iterations}{Number of iterations to use.}

\item{clusters}{Number of clusters for parallel computing.}

\item{seed}{Seed number to use for parallel cluster}

\item{k}{gam max knots}
}
\description{
Run bootstrap sampling for N iterations
}
