% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_for_betareg.R
\name{scale_for_betareg}
\alias{scale_for_betareg}
\title{Transform [0,1] bounded variable for beta regression}
\usage{
scale_for_betareg(x, method = "constant")
}
\arguments{
\item{x}{Numeric vector}

\item{method}{"constant" adds/subtracts 1e-7 from 0/1; "sv" uses the transformation from Smithson and Verkulien (2006)}
}
\description{
Transform 0 and 1 values to meet conditions for beta regression.
}
\references{
Smithson, M. & Verkuilen, J. A better lemon squeezer? Maximum-likelihood regression with beta-distributed dependent variables. Psychol. Methods 11, 54–71 (2006). DOI: 10.1037/1082-989X.11.1.54
}
