% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled_distribution.R
\name{scaled_distribution}
\alias{scaled_distribution}
\title{Generate a normalized vector from normal, skew-normal or bimodal distributions}
\usage{
scaled_distribution(size, demographic_comp_pars = NULL)
}
\arguments{
\item{size}{Numeric vector of sizes}

\item{demographic_comp_pars}{A list containing values for the distribution}
}
\description{
Generate a normalized vector from normal, skew-normal or bimodal distributions
}
