% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccal_cv.R
\name{sccal_cv}
\alias{sccal_cv}
\title{Selectivity conditioned on catch-at-length (SCCAL) cross validation}
\usage{
sccal_cv(
  count1,
  count2,
  effort1,
  effort2,
  size1,
  size2,
  block1,
  block2,
  treatment_name1,
  treatment_name2,
  k = NULL,
  n_cores = 1
)
}
\arguments{
\item{count1}{Numeric vector of catch-at-size/age for gear #1}

\item{count2}{Numeric vector of catch-at-size/age for gear #2}

\item{effort1}{Numeric vector of effort for for gear #1}

\item{effort2}{Numeric vector of effort for for gear #2}

\item{size1}{Numeric vector of sizes or ages for gear #2}

\item{block1}{Sample block (i.e. paired sample) for gear #1}

\item{block2}{Sample block (i.e. paired sample) for gear #2}

\item{k}{k to use for GAMs. Automatically set to the minimum of 8 or 3 less than the number of unique values in size.}

\item{n_cores}{Number of cores to use for parallel processing.}

\item{treatment1_name}{1L vector for the name of the first treatment}

\item{treatment2_name}{1L vector for the name of the second treatment}
}
\value{
Returns a data.frame with predicted catch for Poisson, negative binomial, and Tweedie models.
}
\description{
Block-level cross validation for selectivity condition on catch-at-length Poisson, negative binomial, and Tweedie generalized additive models.
}
