% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectivity_at_size.R
\name{selectivity_at_size}
\alias{selectivity_at_size}
\title{Selectivity at size}
\usage{
selectivity_at_size(size, selectivity_opts)
}
\arguments{
\item{size}{Vector of fork sizes}

\item{selectivity_opts}{List containing parameters for selectivity functions. Current support for double normal (pattern 24 in SS3) or logistic (pattern 1 in SS3).}
}
\description{
Selectivity-at-size for dome-shaped or asymptotic selectivity functions.
}
