% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_paired_sample.R
\name{simulate_paired_sample}
\alias{simulate_paired_sample}
\title{Simulate catch for paired sample between two gears/methods}
\usage{
simulate_paired_sample(
  size,
  abundance,
  availability,
  demographic_comp_pars,
  gear_q1,
  gear_q2,
  effort1,
  effort2,
  selectivity_opts1,
  selectivity_opts2,
  return_vars = TRUE
)
}
\arguments{
\item{size}{Sizes as a numeric vector}

\item{abundance}{Numbers-at-size as a numeric vector}

\item{availability}{Proportion of the total population that is within the sampled area.}

\item{demographic_comp_pars}{Local demographic composition parameters that are used to scale abundance as a list. Required named elements of the list vary depending on the distributional assumption (see ?scaled_distribution.).}

\item{gear_q1}{Gear efficiency for gear/method #1 as a 1L numeric vector}

\item{gear_q2}{Same as above for gear/method #2.}

\item{effort1}{Effort for gear/method #1 as a 1L numeric vector}

\item{effort2}{Same as above for gear/method #2.}

\item{selectivity_opts1}{Same as above for gear/method #2.}

\item{return_vars}{If TRUE, returns a list containing variable values used for the simulation. Otherwise, only returns the samples.}
}
\value{
A list containing size samples for both gears.
}
\description{
Simulate catch-at-size for paired sample between two gears/methods based on numbers-at-size, availability, local demographic structure, gear efficiency, effort (e.g. area swept), and selectivity.
}
\examples{
# A vector of sizes
size <- 10:55

# A vector of abundance-at-size
abundance = round(dnorm(size, mean = 35, sd = 10) * 1e5 * (1-rnorm(length(size), mean = 0, sd = 0.1)))

# Proportion of the total population that is available to a haul ----
availability = 0.001

demographic_comp_pars = list(distribution = "normal", mean = 30, sd = 10)

# Gear efficiency ----
gear_q1 = 1
gear_q2 = 0.8

# Effort for the treatments ----
effort1 = 0.5
effort2 = 1

# Options for the selectivity function ----
selectivity_opts1 = list(type = "asymptotic",
                         begin_top = 35,
                         ln_sd1 = 10)

selectivity_opts2 = list(type = "asymptotic",
                         begin_top = 35,
                         ln_sd1 = 10)

sample1 <- simulate_paired_sample(size = size, 
                                  abundance = abundance, 
                                  availability = availability, 
                                  demographic_comp_pars = demographic_comp_pars, 
                                  gear_q1 = gear_q1, 
                                  gear_q2 = gear_q2, 
                                  effort1 = effort1, 
                                  effort2 = effort2, 
                                  selectivity_opts1 = selectivity_opts1, 
                                  selectivity_opts2 = selectivity_opts2)
}
