% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes_labels.R
\name{species_code_label}
\alias{species_code_label}
\title{Length labels and size breaks for each species}
\usage{
species_code_label(x, type = "axis_label", make_factor = FALSE)
}
\arguments{
\item{x}{RACEBASE species code as a numeric vector (e.g. 21740 for pollock)}

\item{type}{Type of value to return, either "axis_label", "common_name" or "bin_width"}

\item{make_factor}{For common_name, should the common name be returned as an ordered factor}
}
\description{
Length labels and size breaks for each species
}
