% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_codes}
\alias{species_codes}
\title{AFSC RACEBASE Species Codes}
\format{
An object of class 'data.frame'
}
\usage{
data(species_codes)
}
\description{
RACEBASE species codes and common names as of November 8, 2023.
}
\keyword{datasets}
