% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sratio_predict_count.R
\name{sratio_predict_count}
\alias{sratio_predict_count}
\title{Predict Count Based on Selectivity Ratio}
\usage{
sratio_predict_count(
  est_count,
  count1,
  count2,
  sampling_factor1 = 1,
  sampling_factor2 = 1,
  effort1 = 1,
  effort2 = 1,
  s12
)
}
\arguments{
\item{est_count}{Character indicating whether to estimate `"count1"` or `"count2"`.}

\item{count1}{Numeric vector of observed counts of size/age in group 1.}

\item{count2}{Numeric vector of observed counts of size/age in group 2.}

\item{sampling_factor1}{Numeric vector of sampling factors for group 1 (default = 1).}

\item{sampling_factor2}{Sampling factor for group 2 (default = 1).}

\item{effort1}{Effort for group 1 (default = 1).}

\item{effort2}{Effort for group 2 (default = 1).}

\item{s12}{Numeric vector of selectivity ratios between group 1 and group 2, where s12 = selectivity1/selectivity2. Values of `0` are automatically adjusted to a small number (`1e-12`) to avoid dividing by zero.}
}
\value{
Predicted count for the specified group (`count1` or `count2`), rounded to six decimal places.
}
\description{
Calculate the predicted count for one group based on the observed count of the other group, adjusted by the selectivity ratio (`s12`), sampling factor, and effort. If `est_count` is `"count1"`, it estimates the value of `count1` based on `count2` and vice versa.
}
