% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_catch_bootstrap.R
\name{two_stage_bootstrap}
\alias{two_stage_bootstrap}
\title{Two-stage bootstrap draw with resampling}
\usage{
two_stage_bootstrap(x, group, frequency = NULL, draws)
}
\arguments{
\item{x}{A vector of the variable to be sampled.}

\item{group}{A vector of grouping variables for x.}

\item{frequency}{A numeric vector giving the frequency of x. Used for probabalistic sample draws. Default = NULL assigns equal sampling probability to each item x.}

\item{draws}{Number of group samples to draw}
}
\value{
A list containing a list of samples and associated group from which the samples were drawn.
}
\description{
Two-stage bootstrap draw with resampling
}
