% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weighted_mean}
\alias{weighted_mean}
\title{Calculate weighted mean}
\usage{
weighted_mean(x, w, na_rm = FALSE)
}
\arguments{
\item{x}{A numeric vector of values for which the weighted mean is to be calculated.}

\item{w}{A numeric vector of weights corresponding to each value.}

\item{na_rm}{Should NA's be ignored?}
}
\value{
The weighted mean of the values.
}
\description{
Calculate the weighted mean of a numeric vector, given corresponding weights.
}
\examples{
x <- c(1, 2, 3, 4)
w <- c(0.1, 0.2, 0.3, 0.4)
weighted_mean(x, w)
}
