% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectivity_ratio.R
\name{selectivity_ratio}
\alias{selectivity_ratio}
\title{Calculate selectivity catch comparison rate and selectivity ratio}
\usage{
selectivity_ratio(count1, count2, effort1 = NULL, effort2 = NULL)
}
\arguments{
\item{count1}{Count for gear #1}

\item{count2}{Count for gear #2}

\item{effort1}{Effort for gear #1}

\item{effort2}{Effort for gear #2}
}
\description{
Calculate selectivity catch comparison rate and selectivity ratio
}
