% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sratio_cv.R
\name{sratio_cv}
\alias{sratio_cv}
\title{Selectivity ratio cross-validation}
\usage{
sratio_cv(
  count1,
  count2,
  effort1,
  effort2,
  size,
  block,
  k = NULL,
  scale_method = "sv",
  n_cores = 1
)
}
\arguments{
\item{count1}{Numeric vector of catch-at-size/age for gear #1}

\item{count2}{Numeric vector of catch-at-size/age for gear #2}

\item{effort1}{Numeric vector of effort for for gear #1}

\item{effort2}{Numeric vector of effort for for gear #2}

\item{size}{Numeric vector of sizes or ages}

\item{block}{Treatment block (i.e. paired sample)}

\item{k}{k to use for GAMs. Automatically set to the minimum of 8 or 3 less than the number of unique values in size.}

\item{scale_method}{Method to use for scaling the catch comparison rate. "none" = no scaling. For other options see ?scale_for_betareg}

\item{n_cores}{Number of cores to use for parallel processing.}
}
\description{
Block-level selectivity ratio cross validation
}
