% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_stage_bootstrap.R
\name{two_stage_bootstrap}
\alias{two_stage_bootstrap}
\title{Two-stage bootstrap from paired samples}
\usage{
two_stage_bootstrap(
  count1,
  count2,
  size1,
  size2,
  block1,
  block2,
  n_draws,
  seed,
  treatment_name1 = 1,
  treatment_name2 = 2
)
}
\arguments{
\item{count1}{Numeric vector of catch-at-size/age for gear #1}

\item{count2}{Numeric vector of catch-at-size/age for gear #2}

\item{size1}{Numeric vector of sizes or ages for gear #1}

\item{size2}{Numeric vector of sizes or ages for gear #2}

\item{block1}{Sample block for gear #1}

\item{block2}{Sample block for gear #2}

\item{n_draws}{Number of samples to draw.}

\item{seed}{random number generator seed}

\item{treatment_name1}{Name of the treatment for the first gear as a number or character.}

\item{treatment_name2}{Name of the treatment for the second gear as a number or character.}
}
\description{
Two-stage bootstrap resampling (by block/pair then individual within each pair), where samples are drawn from paired treatments.
}
